Revert "Adding copyrights and moving remaining 2 object classes as engine objects"
This reverts commit 96b48e62397eb0227e7c20ba5eba236d691d7813.
