+++ Node2Code_Toolkit.sln
@@ -3,6 +3,8 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.757
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Node2Code_oM", "Node2Code_oM\Node2Code_oM.csproj", "{E35FD3CF-5B58-49C2-B44A-E97F656F009B}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Node2Code_Engine", "Node2Code_Engine\Node2Code_Engine.csproj", "{3FA232F2-1078-4682-B402-CE20B9665856}"
 EndProject
 Global
@@ -11,6 +13,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E35FD3CF-5B58-49C2-B44A-E97F656F009B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E35FD3CF-5B58-49C2-B44A-E97F656F009B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E35FD3CF-5B58-49C2-B44A-E97F656F009B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E35FD3CF-5B58-49C2-B44A-E97F656F009B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3FA232F2-1078-4682-B402-CE20B9665856}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3FA232F2-1078-4682-B402-CE20B9665856}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3FA232F2-1078-4682-B402-CE20B9665856}.Release|Any CPU.ActiveCfg = Release|Any CPU
