+++ Node2Code_oM/Node2Code_oM.csproj
@@ -5,11 +5,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{3FA232F2-1078-4682-B402-CE20B9665856}</ProjectGuid>
+    <ProjectGuid>{E35FD3CF-5B58-49C2-B44A-E97F656F009B}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>BH.Engine.Node2Code</RootNamespace>
-    <AssemblyName>Node2Code_Engine</AssemblyName>
+    <RootNamespace>BH.oM.Node2Code</RootNamespace>
+    <AssemblyName>Node2Code_oM</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -39,29 +39,14 @@
       <HintPath>..\..\BHoM\Build\BHoM.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="Library_Engine">
-      <HintPath>..\..\BHoM_Engine\Build\Library_Engine.dll</HintPath>
-    </Reference>
     <Reference Include="Microsoft.CodeAnalysis, Version=3.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.CodeAnalysis.Common.3.4.0-beta3-final\lib\netstandard2.0\Microsoft.CodeAnalysis.dll</HintPath>
+      <HintPath>..\packages\Microsoft.CodeAnalysis.Common.3.4.0-beta4-final\lib\netstandard2.0\Microsoft.CodeAnalysis.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.CodeAnalysis.CSharp, Version=3.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.CodeAnalysis.CSharp.3.4.0-beta3-final\lib\netstandard2.0\Microsoft.CodeAnalysis.CSharp.dll</HintPath>
-    </Reference>
-    <Reference Include="Programming_Engine">
-      <HintPath>..\..\BHoM_Engine\Build\Programming_Engine.dll</HintPath>
-      <Private>False</Private>
+      <HintPath>..\packages\Microsoft.CodeAnalysis.CSharp.3.4.0-beta4-final\lib\netstandard2.0\Microsoft.CodeAnalysis.CSharp.dll</HintPath>
     </Reference>
     <Reference Include="Programming_oM">
       <HintPath>..\..\BHoM\Build\Programming_oM.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="Reflection_Engine">
-      <HintPath>..\..\BHoM_Engine\Build\Reflection_Engine.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="Reflection_oM">
-      <HintPath>..\..\BHoM\Build\Reflection_oM.dll</HintPath>
       <Private>False</Private>
     </Reference>
     <Reference Include="System" />
@@ -99,40 +84,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Compute\ApplyLoops.cs" />
-    <Compile Include="Compute\Variables.cs" />
-    <Compile Include="Compute\NodeSequence.cs" />
-    <Compile Include="Compute\ApplyGroups.cs" />
-    <Compile Include="Convert\ToCSharpText.cs" />
-    <Compile Include="Convert\ToCSharpSyntaxNode.cs" />
-    <Compile Include="Create\LoopStatement.cs" />
-    <Compile Include="Create\Variable.cs" />
-    <Compile Include="Create\ReturnStatement.cs" />
-    <Compile Include="Objects\BodyState.cs" />
-    <Compile Include="Objects\NodeState.cs" />
-    <Compile Include="Objects\ReceiverState.cs" />
-    <Compile Include="Query\OutputVariable.cs" />
-    <Compile Include="Query\Children.cs" />
-    <Compile Include="Query\Expression.cs" />
-    <Compile Include="Query\SyntaxToken.cs" />
-    <Compile Include="Query\DepthDifference.cs" />
-    <Compile Include="Query\Statements.cs" />
-    <Compile Include="Modify\AddComments.cs" />
-    <Compile Include="Modify\Order.cs" />
+    <Compile Include="Others\StatementInfo.cs" />
+    <Compile Include="Others\Variable.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Query\InlineExpression.cs" />
-    <Compile Include="Query\MethodName.cs" />
-    <Compile Include="Query\Body.cs" />
-    <Compile Include="Query\ArgumentValue.cs" />
-    <Compile Include="Query\ReturnType.cs" />
-    <Compile Include="Query\MethodParameters.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Node2Code_oM\Node2Code_oM.csproj">
-      <Project>{e35fd3cf-5b58-49c2-b44a-e97f656f009b}</Project>
-      <Name>Node2Code_oM</Name>
-      <Private>False</Private>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
