+++ Node2Code_oM/packages.config
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="2.9.6" targetFramework="net461" developmentDependency="true" />
+  <package id="Microsoft.CodeAnalysis.Common" version="3.4.0-beta4-final" targetFramework="net461" />
+  <package id="Microsoft.CodeAnalysis.CSharp" version="3.4.0-beta4-final" targetFramework="net461" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net461" />
+  <package id="System.Collections.Immutable" version="1.5.0" targetFramework="net461" />
+  <package id="System.Memory" version="4.5.3" targetFramework="net461" />
+  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net461" />
+  <package id="System.Reflection.Metadata" version="1.6.0" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net461" />
+  <package id="System.Text.Encoding.CodePages" version="4.5.1" targetFramework="net461" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.3" targetFramework="net461" />
+</packages>
